requirements.txt
pipecat-ai[google,websocket]
fastapi
uvicorn
python-dotenv
aiohttp
